Edit styles, refactor code
