+++ src/pages/editSpecialtyPage/editSpecialty.module.scss
@@ -1,10 +1,4 @@
 @import '../../styles/index.scss';
-.mainInfo {
-  background-color: #bad0e5;
-  border-radius: 0.5rem;
-  padding: 20px;
-  width: 1050px;
-}
 
 .customInput {
   border-radius: 0.5rem;
@@ -20,6 +14,12 @@
 }
 
 .editSpecialty {
+  .mainInfo {
+    background-color: #bad0e5;
+    border-radius: 0.5rem;
+    padding: 1.25rem;
+    width: 65.625rem;
+  }
   display: flex;
   min-height: 100vh;
   .mainContent {
@@ -33,20 +33,20 @@
     text-align: center;
   }
 
-  h1 {
+  .title {
     @extend .heading;
     font-size: $font_30px;
-    margin-top: 50px;
+    margin-top: 3.125rem;
   }
 
-  h2 {
-    @extend .heading, h1;
-    border-bottom: 2px solid #000;
-    margin-bottom: 30px;
+  .infoTitle {
+    @extend .heading, .title;
+    border-bottom: 0.125rem solid #000;
+    margin-bottom: 1.875rem;
     width: max-content;
   }
 
-  h3 {
+  .subtitle {
     @extend .heading;
     font-size: $font_18px;
   }
